requirements.txt
torch
transformers
sentence-transformers
weaviate-client
sqlglot
streamlit
pandas
numpy
clickhouse-connect
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-weaviate
google-generativeai>=0.3.0